(housekeeping) cleanup, reorganization and presentation
